test/src/requirements.txt
Django==1.10.1
Pillow==3.4.1
psycopg2==2.6.2
django-environ==0.4.0
django-extensions
ipdb
pytest
pytest-django
django-redis==4.5.0
redis>=2.10.5
